% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegasos.r
\name{pegasos}
\alias{pegasos}
\title{pegasos}
\usage{
pegasos(x, y, intercept = TRUE, k = 5, lambda = 0.1, niter = 1e+05)
}
\arguments{
\item{x}{The features matrix.}

\item{y}{The 2-class response vector. Note that the groups must be +/- 1.
See \code{recode()}}

\item{intercept}{Should the model include an intercept term?}

\item{k}{TODO, currently ignored}

\item{lambda}{Regularization parameter.}

\item{niter}{The number of iterations.}
}
\value{
An object of class pegasosSVM.
}
\description{
If \code{intercept=TRUE}, the data is not literally modified to include a
column of ones.  However, this is conceptually what we do.
}
\details{
Pegasos SVM solver.
}
\examples{
\dontrun{
library(ssvm)

y <- recode(iris$Species == "setosa")
x <- as.matrix(iris[, -5])

pegasos(x, y)
}

}
\author{
Drew Schmidt
}
\references{
Shalev-Shwartz, Shai, et al. "Pegasos: Primal estimated sub-gradient solver
for svm." Mathematical programming 127.1 (2011): 3-30.
}

