% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.r
\name{recode}
\alias{recode}
\title{recode}
\usage{
recode(response, validate = TRUE)
}
\arguments{
\item{response}{The class variable.}

\item{validate}{Should the response be checked for more than 2 classes? If
so and more than 2 are found, it will throw an error.}
}
\value{
An integer vector of +/- 1's.
}
\description{
Recode a class variable to -1, 1 groups for use with SVM.
}
\examples{
\dontrun{
recode(1:2)
recode(1:3) # two-class only!
}

}

