% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssvm-pkg.r
\docType{package}
\name{ssvm-package}
\alias{ssvm-package}
\title{ssvm: Simple SVM}
\description{
A simple two-class linear SVM implementation using
the Pegasos algorithm.
}
\author{
Drew Schmidt
}
\keyword{package}

